+++ pybis/src/python/pybis/dataset.py
@@ -550,7 +550,7 @@ class DataSet(
 
         fast_download = FastDownload(self.openbis.token, download_url, self.permId, files,
                                      final_destination, create_default_folders, wait_until_finished,
-                                     self.openbis.verify_certificates,
+                                     self.openbis.verify_certificates, self.openbis.get_server_information(),
                                      wished_number_of_streams=4)
         return fast_download.download()
 
@@ -1008,11 +1008,11 @@ class DataSet(
             parent_ids = []
         counter = 5
         for parent_id in parent_ids:
-            param["parentIds"] += {
+            param["parentIds"] += [{
                 "@type": "as.dto.dataset.id.DataSetPermId",
                 "@id": str(counter),
                 "permId": parent_id
-            }
+            }]
             counter += 1
 
         request = {