+++ pybis/src/python/tests/test_dataset.py
@@ -311,3 +311,33 @@ def test_dataset_property_in_isoformat_date(space):
 
     assert len(dataset.p()) == 1
     assert dataset.p[property_type_code] is not None
+
+def test_create_new_dataset_with_parent(space):
+    openbis_instance = space.openbis
+
+    testfile_path = os.path.join(os.path.dirname(__file__), "testdir/testfile")
+
+    dataset = openbis_instance.new_dataset(
+        type="RAW_DATA",
+        experiment="/DEFAULT/DEFAULT/DEFAULT",
+        files=[testfile_path],
+        props={"$name": "some good name"},
+    )
+    dataset.save()
+
+    assert dataset.permId is not None
+    assert dataset.file_list == ["original/testfile"]
+
+    new_dataset = openbis_instance.new_dataset(
+        type="RAW_DATA",
+        experiment="/DEFAULT/DEFAULT/DEFAULT",
+        files=[testfile_path],
+        props={"$name": "some good name"},
+        parents=[dataset.permId]
+    )
+    new_dataset.save()
+
+    assert new_dataset.permId is not None
+    assert new_dataset.file_list == ["original/testfile"]
+    assert new_dataset.parents == [dataset.permId]
+