+++ pybis/src/python/tests/test_fastdownload.py
@@ -1,3 +1,17 @@
+#   Copyright ETH 2018 - 2024 Zürich, Scientific IT Services
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+#
 import binascii
 import json
 import os
@@ -125,6 +139,12 @@ def run_around_tests(base_data):
     yield temp_folder, download_url, streams, perm_id, file
     cleanup(temp_folder)
 
+def get_server_information(major, minor):
+    class ServerInfo:
+        def is_version_greater_than(self, x, y):
+            return x > major or (x == major and y > minor)
+
+    return ServerInfo()
 
 def test_download_fails_after_retries(run_around_tests):
     temp_folder, download_url, streams, perm_id, file = run_around_tests
@@ -136,7 +156,7 @@ def test_download_fails_after_retries(run_around_tests):
     MyServer.next_response['download'] = generate_download_response()
 
     fast_download = FastDownload("", download_url, perm_id, file, str(temp_folder),
-                                 True, True, False, streams)
+                                 True, True, False, get_server_information(3,7), streams)
     try:
         fast_download.download()
         assert False
@@ -162,7 +182,7 @@ def test_download_file(run_around_tests):
     MyServer.next_response['download'] = generate_download_response()
 
     fast_download = FastDownload("", download_url, perm_id, file, str(temp_folder),
-                                 True, True, False, streams)
+                                 True, True, False, get_server_information(3,7), streams)
     fast_download.download()
 
     downloaded_files = [
@@ -201,7 +221,7 @@ def test_download_file_wait_flag_disabled(run_around_tests):
     MyServer.next_response['download'] = generate_download_response()
 
     fast_download = FastDownload("", download_url, perm_id, file, str(temp_folder),
-                                 True, False, False, streams)
+                                 True, False, False, get_server_information(3,7), streams)
     fast_download.download()
 
     # Verify that file has not been downloaded yet
@@ -255,7 +275,7 @@ def test_download_file_starts_with_fail(run_around_tests):
     MyServer.next_response['download'] = generate_download_response()
 
     fast_download = FastDownload("", download_url, perm_id, file, str(temp_folder),
-                                 True, True, False, streams)
+                                 True, True, False, get_server_information(3,7), streams)
     fast_download.download()
 
     downloaded_files = [
@@ -300,7 +320,7 @@ def test_download_fails_after_getting_java_exception(run_around_tests):
     MyServer.next_response['download'] = generate_download_response()
 
     fast_download = FastDownload("", download_url, perm_id, file, str(temp_folder),
-                                 True, True, False, streams)
+                                 True, True, False, get_server_information(3,7), streams)
     try:
         fast_download.download()
         assert False
@@ -334,7 +354,7 @@ def test_download_passes_after_getting_java_exception(run_around_tests):
     MyServer.next_response['download'] = generate_download_response()
 
     fast_download = FastDownload("", download_url, perm_id, file, str(temp_folder),
-                                 True, True, False, streams)
+                                 True, True, False, get_server_information(3,7), streams)
     fast_download.download()
 
     downloaded_files = [
@@ -382,7 +402,7 @@ def test_download_file_payload_failure(run_around_tests):
     MyServer.next_response['download'] = generate_download_response()
 
     fast_download = FastDownload("", download_url, perm_id, file, str(temp_folder),
-                                 True, True, False, streams)
+                                 True, True, False, get_server_information(3,7), streams)
     try:
         fast_download.download()
         assert False